Renamed case folder to cases, renamed case.test to cases.test. Verified generated-types no longer creates cases.types.ts file (#504)
