+++ packages/destination-actions/src/destinations/salesforce/cases/generated-types.ts
